+++ Documentation/Changelog.md
@@ -20,6 +20,7 @@ Release date: not released yet
 - Improved performance of a thread-safe LINQ memoization
 - Fixed nullability annotations for `MathEx.Min` and `MathEx.Max` functions
 - Fixed nullability annotations for `LazyInitializerEx` class
+- Fixed issue with UTF-8 BOM encoding returned by `CommandLine.OemEncoding` on Windows
 - Fixed thread safety issues that could occur on architectures with weaker memory models
 
 ### Gapotchenko.FX 2020.1